refactor: extract GraphValidator from MatchCtxt
CFG and DDG validation logic moved to matches/graph.rs.
GraphValidator is independently constructable from MatchContext + Matching,
separating structural constraint checking from backtracking search.
Reduces matches/mod.rs by ~237 lines.
